Update pattern tests
Fixup tests to remove mocking and test basic functionality.
